Refactor and reorganize art, prefabs, and scripts
Moved and renamed various art, prefab, and script assets to new folder structures for improved organization. Added new UI, audio, and weapon assets. Replaced old deck and UI handlers with updated versions. Updated input actions, GameScene, and card skill scriptable objects. Removed obsolete VFX, sound, and UI handler scripts.

+++ Assets/_game/Scripts/Character/PlayerCollector.cs
@@ -8,7 +8,7 @@ namespace Littale {
     [RequireComponent(typeof(Collider2D))]
     public class PlayerCollector : MonoBehaviour {
 
-        public UnityEvent<float> OnCoinCollected;
+        public UnityEvent<int> OnCoinCollected;
 
         PlayerStats player;
         CircleCollider2D detector;
